clusterloader2: less verbose logging in WaitForPods
The function used to dump information periodically at V(2). When used by
WaitForControlledPods it gets called once per deployment (= pod), which for
10000 pods and created a lot of output.

It's better to just log state changes at V(2) and do the periodic logging at
V(5) when called like that. The relevant state change is when waiting starts
and when and how it completes. That's still quite a lot of log messages (two
for each pod), but at V(2) that's okay and allows monitoring progress.

When called once for several pods as in the WaitForRunningPods measurement it
makes sense to have the status dumps at V(2).